Move context to separate file
